The check to see if we were out of available targets for a client was occurring too late, leading to a null pointer exception if it actually happened.  Moved the check up a bit.
Modified the delayed initialization process so that it is no longer
timing dependant.  Previously, Distributor would call startThread().
Then the first thing the thread would do in run() is finish the
initialization.  However, if the JVM took a while to actually start the
thread, Distributor may have reached a point where it expected our
initialization to be complete before it actually was.  So now
Distributor calls finishInitialization(), which does so and then starts
the thread.  No more timing dependancy.

Removed a test that checked to see if a connection returned by
checkForCompletedConnections() was actually connected.  This test was
needed when I first wrote checkForCompletedConnections(), but I long ago
fixed it so that checkForCompletedConnections() always returns
connections that were successful.  Failed connections are handled
seperately by checkForFailedConnections().