Replaced the connections and connectStartTime maps with a pendingConnections map, which stores PendingConnectionState objects as values to track all of the state associated with a pending connection. Added PendingConnectionState inner class.
Modified the delayed initialization process so that it is no longer
timing dependant.  Previously, Distributor would call startThread().
Then the first thing the thread would do in run() is finish the
initialization.  However, if the JVM took a while to actually start the
thread, Distributor may have reached a point where it expected our
initialization to be complete before it actually was.  So now
Distributor calls finishInitialization(), which does so and then starts
the thread.  No more timing dependancy as the initialization is
finished before finishInitialization() returns.  The thread may not be
started yet, but that shouldn't be a problem as pending connections will
just queue up until the thread starts.

Renamed the server variable in initiateConnection() to connToServer to
better reflect what it represents.

Fixed a bug in checkForFailedConnections where connections which had
timed out were removed from the connectStartTime map (now part of
pendingConnections) without going through the iterator, leading to a
ConcurrentModificationException.
In the same place, previously the selection key for that timed out
connection was not canceled, leading to a mess when it came up in select
because the associated entries had been dropped from the
connections/connectStartTime maps.  The selection key is now stored in
the PendingConnectionState object and canceled if the connection times
out.