Added DelayedMover class to handle data that we can't immediately write to the destination.  It runs in its own thread and selects waiting for the destinations to become writeable.  Previously DataMover would just wait until it could write the data, which meant it could be hung indefinately by an unresponsive destination.  Added associated addToReactivateList() and processReactivateList() methods to DataMover. Added call to processReactivateList() in DataMover's run method.
moveData() method now returns a boolean indicating if it was successful
in writing all of the data to the destination.  If moveData() returns
false (indicating that it had to hand some data off to DelayedMover), we
don't try to read more data from that source channel until DelayedMover
is finished and hands the channel back to DataMover via
addToReactivateList().

Added closeConnection() method for closing a connection.  This
consolidated some code that was duplicated in several places.

Added dumpState() method for cleaning up after we close a connection.
This consolidated the cleanup code which was duplicated in several
places.